Split some code into smaller files
